repoman: Move the large depency checks loop to a new plugin ProfileDependsChecks class
